move rc_type to shared
