+++ src/shared/shared-types.d.ts
@@ -33,3 +33,19 @@ export interface AppState {
   saved: LocalSettings;
 }
 
+export interface RC_Config {
+  'log-debug': boolean
+  'log-to-console': boolean
+  'machine-readable-stacktrace': boolean
+}
+
+import { App } from 'electron'
+import { LocaleData } from '../shared/localize'
+
+export interface ExtendedApp extends App {
+  rc: RC_Config;
+  isQuitting: boolean;
+  ipcReady: boolean;
+  localeData?: LocaleData;
+  state?: AppState;
+}
